Failover mechanism mod move to its own source file
